% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_oceo_gui.R
\name{amlr_oceo_gui}
\alias{amlr_oceo_gui}
\title{Open amlrOceo Shiny app}
\usage{
amlr_oceo_gui(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical with default of \code{TRUE};
passed to \code{launch.browser} argument of \code{\link[shiny]{runApp}}}
}
\description{
Open the amlrOceo R Shiny application
}
\examples{
if (interactive()) amlr_oceo_gui(launch.browser = TRUE)

}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
