% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrOceo-package.R
\docType{package}
\name{amlrOceo-package}
\alias{amlrOceo-package}
\alias{amlrOceo}
\title{AMLR Krill and Oceanographic Program data processing and analysis}
\description{
amlrOceo: A to to analyze and visualize data from the AMLR_OCEO database
}
\details{
TODO: This package contains a Shiny app that connects to the AMLR_PINNIPEDS database,
and allows users to analyze and visualize data from said database.
This can be done manually in R, or using the built-in R Shiny app opened through todo
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
