% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_filtered_volume.R
\name{calculate_volume_filtered}
\alias{calculate_volume_filtered}
\alias{calculate_filtered_volume}
\alias{calculate_filtered_volume_tbl}
\title{Calculate filtered volumes for nets}
\usage{
calculate_filtered_volume(x)

calculate_filtered_volume_tbl(
  net,
  net.station.header,
  amlr.station.header,
  tucker.header.data
)
}
\arguments{
\item{x}{a data frame the necessary columns; see Details}

\item{net}{data frame;}

\item{net.station.header}{data frame;}

\item{amlr.station.header}{data frame;}

\item{tucker.header.data}{data frame;}
}
\value{
A data frame...
}
\description{
Recreates vCalculate_Filtered_Volume
}
\details{
calculate_filtered_volume takes a joined data frame as the input
calculate_filtered_volume_tbl takes the table data frames as inputs
}
\examples{
calculate_filtered_volume(data.frame(
  net_id = 1:7, net_station_header_id = 1, amlr_station_header_id = 1,
  cruise = "AMLR_test", station = "St1", 
  net_type = c("I", "I", "B", "B", "T", "T", "VRS"),
  net_num = 1:7, 
  final_reading = rep(1000, 7), 
  initial_reading = rep(1, 7),
  depth_fished = as.integer(rep(170, 7)), 
  swept_volume = c(NA, NA, NA, NA, 50, 50, NA), 
  tot_dist = c(NA, NA, NA, NA, 1000, 1000, NA), 
  max_depth = as.integer(c(NA, NA, NA, NA, 179, 179, NA))
))

\dontrun{
  calculate_filtered_volume_tbl(
    oceo_net(net), 
    oceo_net_station_header(net.station.header), 
    oceo_amlr_station_header(amlr.station.header), 
    oceo_tucker_header_data(tucker.header.data)
  )
}

}
