% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krill_length_frequency.R
\name{krill_length_frequency}
\alias{krill_length_frequency}
\title{Calculate krill length frequencies}
\usage{
krill_length_frequency(
  net,
  net.station.header,
  amlr.station.header,
  tucker.header.data = NULL
)
}
\arguments{
\item{net}{data frame;}

\item{net.station.header}{data frame;}

\item{amlr.station.header}{data frame;}

\item{tucker.header.data}{data frame;}
}
\value{
A data frame...
}
\description{
Calculate krill length frequencies from provided data
}
\details{
stuff
}
\examples{
\dontrun{
  krill_length_frequency(
    net, net.station.header, amlr.station.header, tucker.header.data
  )
}

}
