% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cruise_info.R
\name{mod_cruise_info}
\alias{mod_cruise_info}
\alias{mod_cruise_info_ui}
\alias{mod_cruise_info_server}
\title{Shiny module for OCEO cruise information}
\usage{
mod_cruise_info_ui(id, col.width = 7)

mod_cruise_info_server(id, pool)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny]{NS}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{pool}{reactive; a DBI database connection pool
Intended to be the output of \code{\link[amlrDatabases]{mod_database_server}}}
}
\value{
A data frame with the contents of the AMLR_STATION_HEADER table,
with some columns renamed
}
\description{
Shiny module for OCEO cruise information
}
