% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_headers.R
\name{mod_headers}
\alias{mod_headers}
\alias{mod_headers_ui}
\alias{mod_headers_server}
\title{Shiny module for OCEO header information}
\usage{
mod_headers_ui(id, col.width = 7)

mod_headers_server(id, pool)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny]{NS}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{pool}{reactive; a DBI database connection pool
Intended to be the output of \code{\link[amlrDatabases]{mod_database_server}}}
}
\value{
A list of reactive data frames from HEADER tables:
\enumerate{
\item \code{amlr_station_header}:
a reactive data frame with the contents of the AMLR_STATION_HEADER table,
with some columns renamed. The column names of this output data frame are:
amlr_station_header_id, cruise_station, cruise, station, leg, ship, area,
bottom_depth, created_dt, ice, restrictedUse, gssmu
\item \code{net_station_header}:
a reactive data frame with the contents of the NET_STATION_HEADER table,
with some columns renamed. The column names of this output data frame are:
tbd
}
}
\description{
Shiny module for OCEO header information
}
