% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceo_table_process.R
\name{oceo_table_process}
\alias{oceo_table_process}
\alias{oceo_amlr_station_header}
\alias{oceo_net_station_header}
\alias{oceo_net}
\alias{oceo_tucker_header_data}
\title{Renaming functions}
\usage{
oceo_amlr_station_header(x)

oceo_net_station_header(x)

oceo_net(x)

oceo_tucker_header_data(x)
}
\arguments{
\item{x}{data frame, a table from the AMLR_OCEO database}
}
\value{
The data frame \code{x}, with columns processed as described below:
\enumerate{
\item \code{oceo_amlr_station_header}. Output column names:
amlr_station_header_id, cruise_station, cruise, station, leg, ship, area,
bottom_depth, created_dt, ice, restrictedUse, gssmu
\item \code{oce_net_station_header}.
time_local is converted to a factor, and date_tow is converted to a Date object.
Output column names: tbd
}
}
\description{
Functions that rename columns in data frames extracted from data frames int eh AMLR_OCEO database
}
\details{
The AMLR_OCEO database has a number of views, stored procedures, and other items
that make it difficult to change the names of fields in the table.
Thus, the purpose of these functions is to have a single place where the
columns are renamed to something more consistent and R-friendly.
These functions also make it so that only one place needs to be changed
if the table column names do change.
}
\examples{
\dontrun{
  tbl(con, "table_name") \%>\% 
    collect() \%>\%
    oceo_process_amlr_station_header()
}
}
