% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_krill_length_frequency.R, R/shiny_modules.R
\name{shiny_modules}
\alias{shiny_modules}
\alias{mod_krill_length_frequency_ui}
\alias{mod_krill_length_frequency_server}
\title{Modules for the different tabs of the amlrOceo Shiny app}
\usage{
mod_krill_length_frequency_ui(id)

mod_krill_length_frequency_server(id, pool, amlr.station.header)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{pool}{reactive; a DBI database connection pool.
Intended to be the output of \code{\link[amlrDatabases]{mod_database_server}}}

\item{amlr.station.header}{reactive; the output of \code{\link{mod_cruise_info_server}}}
}
\description{
Modules for the different tabs of the amlrOceo Shiny app
}
