% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_filtered_volume.R
\name{calculate_filtered_volume}
\alias{calculate_filtered_volume}
\title{Calculate filtered volumes for nets}
\usage{
calculate_filtered_volume(
  net,
  net.station.header,
  amlr.station.header,
  tucker.header.data
)
}
\arguments{
\item{net}{data frame;}

\item{net.station.header}{data frame;}

\item{amlr.station.header}{data frame;}

\item{tucker.header.data}{data frame;}
}
\value{
A data frame...
}
\description{
Recreates vCalculate_Filtered_Volume
}
\details{
stuff
}
\examples{
con <- DBI::dbConnect(odbc::odbc(), 
  Driver = "SQL Server", #"SQL Server Native Client 11.0", 
  Server = "swc-estrella-s",
  Database = "AMLR_OCEO",
  Trusted_Connection = "True")

stopifnot(require(dplyr))
net                 <- tbl(con, "NET") \%>\% collect()
net.station.header  <- tbl(con, "NET_STATION_HEADER") \%>\% collect()
amlr.station.header <- tbl(con, "AMLR_STATION_HEADER") \%>\% collect()
tucker.header.data  <- tbl(con, "TUCKER_HEADER_DATA") \%>\% collect()

calculate_filtered_volume(net, net.station.header, amlr.station.header, tucker.header.data)

}
