% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cruise_info.R
\name{mod_cruise_info}
\alias{mod_cruise_info}
\alias{mod_cruise_info_ui}
\alias{mod_cruise_info_server}
\title{Shiny module for OCEO cruise information}
\usage{
mod_cruise_info_ui(id, col.width = 7)

mod_cruise_info_server(id, pool)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny]{NS}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{pool}{reactive; a DBI database connection pool, intended to be the output of \code{\link{mod_database_server}}}
}
\value{
\code{mod_season_server} returns a list of ...:
\enumerate{
\item \code{season.df}, the season information data frame and
\item \code{season.id.list}, a list of the ID values from the season information table,
with the 'season_name' values as names
}
}
\description{
Shiny module for OCEO cruise information
}
